pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown_maker"
version = "0.1.0"
description = "A CLI tool to convert Confluence pages to Markdown."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click",
    "requests",
    "PyYAML",
    "beautifulsoup4",
    "markdownify",
    "atlassian-python-api",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
    "ruff",
]

[project.scripts]
markdown-maker = "markdown_maker.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.ruff.format]
quote-style = "double"
